% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rlogo.R
\docType{data}
\name{aaColors}
\alias{aaColors}
\title{A color scheme for amino acids}
\format{An object of class \code{character} of length 20.}
\usage{
aaColors
}
\description{
A color scheme for amino acids
}
\keyword{datasets}
