% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rlogo.R
\docType{data}
\name{aaLetters}
\alias{aaLetters}
\title{single letter amino acid codes}
\format{An object of class \code{character} of length 20.}
\usage{
aaLetters
}
\description{
single letter amino acid codes
}
\keyword{datasets}
