% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rlogo.R
\name{drawLetter}
\alias{drawLetter}
\title{draw letter}
\usage{
drawLetter(letter, x0, y0, x1, y1, col = par("fg"), bgcol = "white")
}
\arguments{
\item{letter}{a single character string of the letter to draw}

\item{x0}{the leftmost x-coordinate of the bounding box}

\item{y0}{the bottom y-coordinate of the bounding box}

\item{x1}{the rightmost x-coordinate of the bounding box}

\item{y1}{the top y-coordinate of the bounding box}

\item{col}{the color in which to draw the letter, defaults to the foreground color}

\item{bgcol}{the background color used to fill cut-out parts of letters}
}
\description{
draws a single letter as a polygon within a given bounding box
}
