% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rlogo.R
\name{drawSimpleLogo}
\alias{drawSimpleLogo}
\title{draw a simple weblogo}
\usage{
drawSimpleLogo(data, col, lttrs = rownames(data))
}
\arguments{
\item{data}{a numerical matrix with rownames corresponding to the letters to be drawn.}

\item{col}{a named vector of colors to be used for each letter.}

\item{lttrs}{the letters corresponding to the rows in the matrix, defaults to the matrix's colnames.}
}
\description{
draws a simple web logo for a matrix of values.
}
