% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rlogo.R
\docType{data}
\name{ncColors}
\alias{ncColors}
\title{A color scheme for nucleotides}
\format{An object of class \code{character} of length 4.}
\usage{
ncColors
}
\description{
A color scheme for nucleotides
}
\keyword{datasets}
