% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rlogo.R
\docType{data}
\name{ncLetters}
\alias{ncLetters}
\title{single letter nucleotide codes}
\format{An object of class \code{character} of length 4.}
\usage{
ncLetters
}
\description{
single letter nucleotide codes
}
\keyword{datasets}
